requirements.txt
beautifulsoup4==4.10.0
discord.py==1.7.3
google_api_python_client==2.38.0
matplotlib==3.5.1
pandas==1.4.1
Pillow==9.0.1
python-dotenv==0.19.2
requests==2.27.1
